communication identity related codes replaced to tests folder from shared
